fix(web-search): clarify web tool fallback behavior
