+++ src/renderer/i18n/locales/de-de.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "Speichern fehlgeschlagen",
   "settings.tool.websearch.errors.zhipu_sync_failed": "Synchronisierung des Zhipu-API-Schlüssels mit der Websuche fehlgeschlagen. Bitte speichern Sie den Schlüssel erneut oder überprüfen Sie die Einstellungen der Websuche.",
   "settings.tool.websearch.fetch_urls_provider": "URL-Abrufanbieter",
-  "settings.tool.websearch.model_tools_preferred.description": "Wenn die modellintegrierte Suche nicht verfügbar ist, wird automatisch zu den oben aufgeführten Diensten gewechselt.",
-  "settings.tool.websearch.model_tools_preferred.label": "Modellintegrierte Suche bevorzugen",
+  "settings.tool.websearch.model_tools_preferred.description": "Modellintegrierte Websuche und URL-Abrufe bevorzugen und bei Nichtverfügbarkeit automatisch zu den oben aufgeführten Diensten wechseln.",
+  "settings.tool.websearch.model_tools_preferred.label": "Modellintegrierte Web-Tools bevorzugen",
   "settings.tool.websearch.provider_description.bocha": "Chinesische KI-Such-API mit Echtzeit-Web- und strukturierten Ergebnissen.",
   "settings.tool.websearch.provider_description.exa": "Neural Search-API für KI-Anwendungen, optimiert für die semantische Webabfrage.",
   "settings.tool.websearch.provider_description.exa_mcp": "Stellt Agenten die Exa-Suche über den Exa-MCP-Server zur Verfügung.",
