+++ src/renderer/i18n/locales/el-gr.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "Η αποθήκευση απέτυχε",
   "settings.tool.websearch.errors.zhipu_sync_failed": "Αποτυχία συγχρονισμού του κλειδιού API του Zhipu με την Αναζήτηση Ιστού. Παρακαλώ αποθηκεύστε ξανά το κλειδί ή ελέγξτε τις ρυθμίσεις Αναζήτησης Ιστού.",
   "settings.tool.websearch.fetch_urls_provider": "Πάροχος λήψης διευθύνσεων URL",
-  "settings.tool.websearch.model_tools_preferred.description": "Όταν η ενσωματωμένη υπηρεσία αναζήτησης του μοντέλου δεν είναι διαθέσιμη, γίνεται αυτόματη μετάβαση στις παραπάνω υπηρεσίες.",
-  "settings.tool.websearch.model_tools_preferred.label": "Προτίμηση ενσωματωμένης αναζήτησης μοντέλου",
+  "settings.tool.websearch.model_tools_preferred.description": "Προτιμά την εγγενή αναζήτηση ιστού και ανάκτηση URL του μοντέλου και μεταβαίνει αυτόματα στις παραπάνω υπηρεσίες όταν δεν είναι διαθέσιμες.",
+  "settings.tool.websearch.model_tools_preferred.label": "Προτίμηση εγγενών εργαλείων ιστού του μοντέλου",
   "settings.tool.websearch.provider_description.bocha": "Κινέζικο API αναζήτησης τεχνητής νοημοσύνης με πραγματικές δυνατότητες στον ιστό και δομημένα αποτελέσματα.",
   "settings.tool.websearch.provider_description.exa": "API νευρωνικής αναζήτησης για εφαρμογές τεχνητής νοημοσύνης, βελτιστοποιημένο για σημασιολογική ανάκτηση από τον ιστό.",
   "settings.tool.websearch.provider_description.exa_mcp": "Κάντε διαθέσιμη την αναζήτηση Exa στους πράκτορες μέσω του διακομιστή Exa MCP.",
