+++ src/renderer/i18n/locales/en-us.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "Save failed",
   "settings.tool.websearch.errors.zhipu_sync_failed": "Failed to sync Zhipu API key to Web Search. Please re-save the key or check Web Search settings.",
   "settings.tool.websearch.fetch_urls_provider": "URL fetch provider",
-  "settings.tool.websearch.model_tools_preferred.description": "Automatically switch to the services above when model-native search is unavailable.",
-  "settings.tool.websearch.model_tools_preferred.label": "Prefer model-native search",
+  "settings.tool.websearch.model_tools_preferred.description": "Prefer model-native web search and URL fetching, with automatic fallback to the services above when unavailable.",
+  "settings.tool.websearch.model_tools_preferred.label": "Prefer model-native web tools",
   "settings.tool.websearch.provider_description.bocha": "Chinese AI search API with real-time web and structured results.",
   "settings.tool.websearch.provider_description.exa": "Neural search API for AI apps, tuned for semantic web retrieval.",
   "settings.tool.websearch.provider_description.exa_mcp": "Expose Exa search to agents through the Exa MCP Server.",
