+++ src/renderer/i18n/locales/es-es.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "Error al guardar",
   "settings.tool.websearch.errors.zhipu_sync_failed": "Error al sincronizar la clave API de Zhipu con Web Search. Vuelve a guardar la clave o verifica la configuración de Web Search.",
   "settings.tool.websearch.fetch_urls_provider": "Proveedor de obtención de URL",
-  "settings.tool.websearch.model_tools_preferred.description": "Cambia automáticamente a los servicios anteriores cuando la búsqueda integrada del modelo no está disponible.",
-  "settings.tool.websearch.model_tools_preferred.label": "Preferir la búsqueda integrada del modelo",
+  "settings.tool.websearch.model_tools_preferred.description": "Prioriza la búsqueda web y la obtención de URL nativas del modelo, y cambia automáticamente a los servicios anteriores cuando no están disponibles.",
+  "settings.tool.websearch.model_tools_preferred.label": "Preferir las herramientas web nativas del modelo",
   "settings.tool.websearch.provider_description.bocha": "API de búsqueda de IA china con resultados web en tiempo real y estructurados.",
   "settings.tool.websearch.provider_description.exa": "API de búsqueda neuronal para aplicaciones de IA, optimizada para la recuperación semántica en la web.",
   "settings.tool.websearch.provider_description.exa_mcp": "Expón la búsqueda de Exa a los agentes a través del Servidor MCP de Exa.",
