+++ src/renderer/i18n/locales/fr-fr.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "Échec de l'enregistrement",
   "settings.tool.websearch.errors.zhipu_sync_failed": "Échec de la synchronisation de la clé API Zhipu avec la Recherche Web. Veuillez ré-enregistrer la clé ou vérifier les paramètres de la Recherche Web.",
   "settings.tool.websearch.fetch_urls_provider": "Fournisseur de récupération d'URL",
-  "settings.tool.websearch.model_tools_preferred.description": "Bascule automatiquement vers les services ci-dessus lorsque la recherche intégrée au modèle n’est pas disponible.",
-  "settings.tool.websearch.model_tools_preferred.label": "Privilégier la recherche intégrée au modèle",
+  "settings.tool.websearch.model_tools_preferred.description": "Privilégie la recherche web et la récupération d’URL natives du modèle, avec basculement automatique vers les services ci-dessus lorsqu’elles sont indisponibles.",
+  "settings.tool.websearch.model_tools_preferred.label": "Privilégier les outils web natifs du modèle",
   "settings.tool.websearch.provider_description.bocha": "API de recherche IA chinoise avec résultats web en temps réel et structurés.",
   "settings.tool.websearch.provider_description.exa": "API de recherche neuronale pour applications IA, optimisée pour la récupération sémantique sur le web.",
   "settings.tool.websearch.provider_description.exa_mcp": "Exposez la recherche Exa aux agents via le serveur MCP Exa.",
