+++ src/renderer/i18n/locales/ja-jp.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "保存に失敗しました",
   "settings.tool.websearch.errors.zhipu_sync_failed": "Zhipu APIキーのWeb Searchへの同期に失敗しました。キーを再保存するか、Web Searchの設定を確認してください。",
   "settings.tool.websearch.fetch_urls_provider": "URL取得プロバイダー",
-  "settings.tool.websearch.model_tools_preferred.description": "モデル内蔵検索サービスが利用できない場合は、上記のサービスに自動的に切り替えます。",
-  "settings.tool.websearch.model_tools_preferred.label": "モデル内蔵検索を優先",
+  "settings.tool.websearch.model_tools_preferred.description": "モデル内蔵の Web 検索と URL 取得を優先し、利用できない場合は上記のサービスへ自動的に切り替えます。",
+  "settings.tool.websearch.model_tools_preferred.label": "モデル内蔵の Web ツールを優先",
   "settings.tool.websearch.provider_description.bocha": "リアルタイムのウェブと構造化された結果を持つ中国のAI検索API",
   "settings.tool.websearch.provider_description.exa": "AIアプリ向けニューラル検索API、セマンティックWeb検索に最適化されています。",
   "settings.tool.websearch.provider_description.exa_mcp": "Exa MCPサーバーを通じて、エージェントにExa検索を公開する。",
