+++ src/renderer/i18n/locales/pt-pt.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "Falha ao guardar",
   "settings.tool.websearch.errors.zhipu_sync_failed": "Falha ao sincronizar a chave de API Zhipu com a Pesquisa na Web. Por favor, guarde novamente a chave ou verifique as configurações da Pesquisa na Web.",
   "settings.tool.websearch.fetch_urls_provider": "Fornecedor de busca de URL",
-  "settings.tool.websearch.model_tools_preferred.description": "Muda automaticamente para os serviços acima quando a pesquisa integrada no modelo não está disponível.",
-  "settings.tool.websearch.model_tools_preferred.label": "Preferir a pesquisa integrada no modelo",
+  "settings.tool.websearch.model_tools_preferred.description": "Dá prioridade à pesquisa Web e à obtenção de URLs nativas do modelo, mudando automaticamente para os serviços acima quando não estão disponíveis.",
+  "settings.tool.websearch.model_tools_preferred.label": "Preferir as ferramentas Web nativas do modelo",
   "settings.tool.websearch.provider_description.bocha": "API de busca chinesa de IA com resultados da web em tempo real e estruturados.",
   "settings.tool.websearch.provider_description.exa": "API de busca neural para aplicações de IA, ajustada para recuperação semântica da web.",
   "settings.tool.websearch.provider_description.exa_mcp": "Exponha a pesquisa Exa aos agentes por meio do Exa MCP Server.",
