+++ src/renderer/i18n/locales/ro-ro.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "Salvarea a eșuat",
   "settings.tool.websearch.errors.zhipu_sync_failed": "Nu s-a reușit sincronizarea cheii API Zhipu cu Căutarea Web. Te rugăm să salvezi din nou cheia sau să verifici setările Căutării Web.",
   "settings.tool.websearch.fetch_urls_provider": "Furnizor de preluare URL",
-  "settings.tool.websearch.model_tools_preferred.description": "Comută automat la serviciile de mai sus când căutarea integrată a modelului nu este disponibilă.",
-  "settings.tool.websearch.model_tools_preferred.label": "Preferă căutarea integrată a modelului",
+  "settings.tool.websearch.model_tools_preferred.description": "Preferă căutarea web și preluarea URL-urilor native ale modelului, comutând automat la serviciile de mai sus când nu sunt disponibile.",
+  "settings.tool.websearch.model_tools_preferred.label": "Preferă instrumentele web native ale modelului",
   "settings.tool.websearch.provider_description.bocha": "API de căutare AI chinezesc cu rezultate web în timp real și structurate.",
   "settings.tool.websearch.provider_description.exa": "API de căutare neuronală pentru aplicații AI, optimizat pentru recuperare semantică pe web.",
   "settings.tool.websearch.provider_description.exa_mcp": "Expune căutarea Exa către agenți prin serverul Exa MCP.",
