+++ src/renderer/i18n/locales/ru-ru.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "Ошибка сохранения",
   "settings.tool.websearch.errors.zhipu_sync_failed": "Не удалось синхронизировать ключ API Zhipu с веб-поиском. Повторно сохраните ключ или проверьте настройки веб-поиска.",
   "settings.tool.websearch.fetch_urls_provider": "Провайдер загрузки URL",
-  "settings.tool.websearch.model_tools_preferred.description": "Автоматически переключаться на указанные выше сервисы, если встроенный поиск модели недоступен.",
-  "settings.tool.websearch.model_tools_preferred.label": "Предпочитать встроенный поиск модели",
+  "settings.tool.websearch.model_tools_preferred.description": "Предпочитать встроенные в модель веб-поиск и получение данных по URL, автоматически переключаясь на указанные выше сервисы, если они недоступны.",
+  "settings.tool.websearch.model_tools_preferred.label": "Предпочитать встроенные веб-инструменты модели",
   "settings.tool.websearch.provider_description.bocha": "Китайский API поиска с ИИ, обеспечивающий веб-результаты в реальном времени и структурированные данные.",
   "settings.tool.websearch.provider_description.exa": "API нейропоиска для ИИ-приложений, оптимизированный для семантического поиска в вебе.",
   "settings.tool.websearch.provider_description.exa_mcp": "Предоставьте агентам доступ к поиску Exa через сервер Exa MCP.",
