+++ src/renderer/i18n/locales/vi-vn.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "Lưu thất bại",
   "settings.tool.websearch.errors.zhipu_sync_failed": "Không thể đồng bộ khóa API Zhipu với Tìm kiếm Web. Vui lòng lưu lại khóa hoặc kiểm tra cài đặt Tìm kiếm Web.",
   "settings.tool.websearch.fetch_urls_provider": "Nhà cung cấp tìm nạp URL",
-  "settings.tool.websearch.model_tools_preferred.description": "Tự động chuyển sang các dịch vụ ở trên khi dịch vụ tìm kiếm tích hợp của mô hình không khả dụng.",
-  "settings.tool.websearch.model_tools_preferred.label": "Ưu tiên tìm kiếm tích hợp của mô hình",
+  "settings.tool.websearch.model_tools_preferred.description": "Ưu tiên tìm kiếm web và truy xuất URL tích hợp của mô hình, tự động chuyển sang các dịch vụ ở trên khi không khả dụng.",
+  "settings.tool.websearch.model_tools_preferred.label": "Ưu tiên công cụ web tích hợp của mô hình",
   "settings.tool.websearch.provider_description.bocha": "API tìm kiếm AI Trung Quốc với kết quả web thời gian thực và có cấu trúc.",
   "settings.tool.websearch.provider_description.exa": "API tìm kiếm thần kinh cho ứng dụng AI, được tối ưu hóa cho việc truy xuất web ngữ nghĩa.",
   "settings.tool.websearch.provider_description.exa_mcp": "Cung cấp tìm kiếm Exa cho Agent qua máy chủ Exa MCP.",
