+++ src/renderer/i18n/locales/zh-cn.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "保存失败",
   "settings.tool.websearch.errors.zhipu_sync_failed": "智谱 API 密钥同步到 Web Search 失败，请重新保存密钥或检查 Web Search 设置。",
   "settings.tool.websearch.fetch_urls_provider": "URL 获取服务商",
-  "settings.tool.websearch.model_tools_preferred.description": "模型内置搜索服务不可用时自动切换到上方服务",
-  "settings.tool.websearch.model_tools_preferred.label": "优先使用模型内置搜索",
+  "settings.tool.websearch.model_tools_preferred.description": "优先使用模型内置的网页搜索和 URL 获取功能；不可用时自动切换到上方服务。",
+  "settings.tool.websearch.model_tools_preferred.label": "优先使用模型内置 Web 工具",
   "settings.tool.websearch.provider_description.bocha": "面向 AI 场景的中文搜索 API，提供实时网页与结构化结果。",
   "settings.tool.websearch.provider_description.exa": "为 AI 应用设计的神经搜索 API，擅长语义检索高质量网页。",
   "settings.tool.websearch.provider_description.exa_mcp": "通过 Exa MCP Server 将 Exa 搜索能力暴露给智能体工具调用。",
