+++ src/renderer/i18n/locales/zh-tw.json
@@ -4880,8 +4880,8 @@
   "settings.tool.websearch.errors.save_failed": "儲存失敗",
   "settings.tool.websearch.errors.zhipu_sync_failed": "智譜 API 金鑰同步到 Web Search 失敗，請重新儲存金鑰或檢查 Web Search 設定。",
   "settings.tool.websearch.fetch_urls_provider": "URL 擷取供應商",
-  "settings.tool.websearch.model_tools_preferred.description": "模型內建搜尋服務不可用時自動切換至上方服務",
-  "settings.tool.websearch.model_tools_preferred.label": "優先使用模型內建搜尋",
+  "settings.tool.websearch.model_tools_preferred.description": "優先使用模型內建的網頁搜尋和 URL 擷取功能；無法使用時自動切換至上方服務。",
+  "settings.tool.websearch.model_tools_preferred.label": "優先使用模型內建 Web 工具",
   "settings.tool.websearch.provider_description.bocha": "面向 AI 場景的中文搜尋 API，提供即時網頁與結構化結果。",
   "settings.tool.websearch.provider_description.exa": "為 AI 應用程式設計的神經搜尋 API，擅長語意檢索高品質網頁。",
   "settings.tool.websearch.provider_description.exa_mcp": "透過 Exa MCP Server 將 Exa 搜尋能力暴露給 Agent 工具呼叫。",
