+++ v2-refactor-temp/docs/breaking-changes/2026-08-31-model-native-web-tools-default.md
@@ -8,16 +8,16 @@ date: 2026-08-31
 
 ## What changed
 
-Web search and URL fetching now use model-native capabilities by default, with an automatic fallback to configured services when the model does not support them or they are unavailable. During upgrade, an existing preference for configured services is reset once to the new model-native default.
+Web search and URL fetching now use model-native capabilities by default, with an automatic fallback to configured services when the model does not support them or they are unavailable. During upgrade, the existing source preference is preserved under the new positive setting.
 
 ## Why this matters to the user
 
-Users who previously preferred configured search services may see supported models handle web search and URL fetching directly after upgrading. Configured services remain available as an automatic fallback.
+New installations prefer model-native web tools. Existing users keep their effective source priority after upgrading, and both choices continue to fall back automatically when the preferred side is unavailable.
 
 ## What the user should do
 
-Nothing — automatic. To prefer configured services again, enable **Prefer configured search services** under Settings → Web Search.
+Nothing — automatic. To prefer configured services, turn off **Prefer model-native web tools** under Settings → Web Search.
 
 ## Notes for release manager
 
-The upgrade reset is one-time only. A user who enables the setting afterward keeps that choice on later launches.
+The preference is migrated once by inverting the legacy value, so its effective routing choice remains unchanged.
